web application using ASP.NET Core MVC that represents a Course Registration program
Create a web application using ASP.NET Core MVC that represents a Course Registration program for a school. The Course Registration program does not need to be connected to a database, but you will need objects on the server to represent Courses, Instructors, and Students.

There should be three views, in addition to a homepage and a main layout with a navigation menu.

The first view is the Instructors view. This view is a page displaying a table with InstructorId, First Name, Last Name, Email address, and the Course that they teach. Each instructor can only teach one course.

The second view is the Students view. This view is a page displaying a table with StudentId, First Name, Last Name, Email address, and Phone number.

Finally, there is a third view for Courses. This view is a page displaying a table with CourseId, Course Number, Course Name, and Description.

[Optional] -- Clicking the course name should open up a modal dialog that displays a table that has all students who have enrolled in that course.

You do not need to implement the ability to add, edit, or delete individual objects but you should include sample data in your application so the views are populated for easy viewing of the available views and features.